Fix bug with `memcmpw(3)`, `memcmpl(3)` and `memcmpq(3)`
Return value might have overflown if word differences were too large. Also: to simplify interoperability with `wmemcmp(3)`, change the return type to `int`

+++ kos/include/wchar.h
@@ -1,4 +1,4 @@
-/* HASH CRC-32:0xc7b6b38c */
+/* HASH CRC-32:0xf45cee19 */
 /* Copyright (c) 2019-2022 Griefer@Work                                       *
  *                                                                            *
  * This software is provided 'as-is', without any express or implied          *
@@ -599,24 +599,24 @@ __NAMESPACE_LOCAL_USING_OR_IMPL(mbsinit, __FORCELOCAL __ATTR_ARTIFICIAL __ATTR_P
 #ifdef __CRT_HAVE_wmemcmp
 /* >> wmemcmp(3) */
 __CDECLARE(__ATTR_PURE __ATTR_WUNUSED __ATTR_INS(1, 3) __ATTR_INS(2, 3) __ATTR_NONNULL((1, 2)),int,__NOTHROW_NCX,wmemcmp,(wchar_t const *__s1, wchar_t const *__s2, size_t __num_chars),(__s1,__s2,__num_chars))
-#elif defined(__CRT_HAVE_memcmpw) && __SIZEOF_WCHAR_T__ == 2 && __SIZEOF_INT__ <= 2
+#elif defined(__CRT_HAVE_memcmpw) && __SIZEOF_WCHAR_T__ == 2
 /* >> wmemcmp(3) */
 __CREDIRECT(__ATTR_PURE __ATTR_WUNUSED __ATTR_INS(1, 3) __ATTR_INS(2, 3) __ATTR_NONNULL((1, 2)),int,__NOTHROW_NCX,wmemcmp,(wchar_t const *__s1, wchar_t const *__s2, size_t __num_chars),memcmpw,(__s1,__s2,__num_chars))
-#elif defined(__CRT_HAVE_memcmpl) && __SIZEOF_WCHAR_T__ == 4 && __SIZEOF_INT__ <= 4
+#elif defined(__CRT_HAVE_memcmpl) && __SIZEOF_WCHAR_T__ == 4
 /* >> wmemcmp(3) */
 __CREDIRECT(__ATTR_PURE __ATTR_WUNUSED __ATTR_INS(1, 3) __ATTR_INS(2, 3) __ATTR_NONNULL((1, 2)),int,__NOTHROW_NCX,wmemcmp,(wchar_t const *__s1, wchar_t const *__s2, size_t __num_chars),memcmpl,(__s1,__s2,__num_chars))
-#elif __SIZEOF_WCHAR_T__ == 2 && __SIZEOF_INT__ <= 2
+#elif __SIZEOF_WCHAR_T__ == 2
 __NAMESPACE_STD_END
 #include <libc/local/string/memcmpw.h>
 __NAMESPACE_STD_BEGIN
 /* >> wmemcmp(3) */
-__FORCELOCAL __ATTR_ARTIFICIAL __ATTR_PURE __ATTR_WUNUSED __ATTR_INS(1, 3) __ATTR_INS(2, 3) __ATTR_NONNULL((1, 2)) int __NOTHROW_NCX(__LIBCCALL wmemcmp)(wchar_t const *__s1, wchar_t const *__s2, size_t __num_chars) { return (int)(__NAMESPACE_LOCAL_SYM __LIBC_LOCAL_NAME(memcmpw))((void const *)__s1, (void const *)__s2, (__SIZE_TYPE__)__num_chars); }
-#elif __SIZEOF_WCHAR_T__ == 4 && __SIZEOF_INT__ <= 4
+__FORCELOCAL __ATTR_ARTIFICIAL __ATTR_PURE __ATTR_WUNUSED __ATTR_INS(1, 3) __ATTR_INS(2, 3) __ATTR_NONNULL((1, 2)) int __NOTHROW_NCX(__LIBCCALL wmemcmp)(wchar_t const *__s1, wchar_t const *__s2, size_t __num_chars) { return (__NAMESPACE_LOCAL_SYM __LIBC_LOCAL_NAME(memcmpw))((void const *)__s1, (void const *)__s2, (__SIZE_TYPE__)__num_chars); }
+#elif __SIZEOF_WCHAR_T__ == 4
 __NAMESPACE_STD_END
 #include <libc/local/string/memcmpl.h>
 __NAMESPACE_STD_BEGIN
 /* >> wmemcmp(3) */
-__FORCELOCAL __ATTR_ARTIFICIAL __ATTR_PURE __ATTR_WUNUSED __ATTR_INS(1, 3) __ATTR_INS(2, 3) __ATTR_NONNULL((1, 2)) int __NOTHROW_NCX(__LIBCCALL wmemcmp)(wchar_t const *__s1, wchar_t const *__s2, size_t __num_chars) { return (int)(__NAMESPACE_LOCAL_SYM __LIBC_LOCAL_NAME(memcmpl))((void const *)__s1, (void const *)__s2, (__SIZE_TYPE__)__num_chars); }
+__FORCELOCAL __ATTR_ARTIFICIAL __ATTR_PURE __ATTR_WUNUSED __ATTR_INS(1, 3) __ATTR_INS(2, 3) __ATTR_NONNULL((1, 2)) int __NOTHROW_NCX(__LIBCCALL wmemcmp)(wchar_t const *__s1, wchar_t const *__s2, size_t __num_chars) { return (__NAMESPACE_LOCAL_SYM __LIBC_LOCAL_NAME(memcmpl))((void const *)__s1, (void const *)__s2, (__SIZE_TYPE__)__num_chars); }
 #else /* ... */
 __NAMESPACE_STD_END
 #include <libc/local/wchar/wmemcmp.h>
